requirements.txt
google-generativeai
pydantic
python-dotenv
pypub3
html2image